Finish renaming of Tablature class to StringData
Renames also the files hoding the class itself:
- libmscore/tablature.cpp => stringdata.cpp
- libmscore/tablature.h => stringdata.h

No actual change in the code, only updated references to stringdata.h in #include's.
