+++ libmscore/stringdata.h
