+++ mscore/fretproperties.cpp
@@ -23,9 +23,9 @@
 #include "libmscore/measure.h"
 #include "libmscore/system.h"
 #include "libmscore/score.h"
+//#include "libmscore/stringdata.h"
 #include "fretcanvas.h"
 #include "preferences.h"
-#include "libmscore/tablature.h"
 #include "libmscore/chord.h"
 #include "libmscore/note.h"
 #include "libmscore/segment.h"
